rename lexicon
