+++ packages/api/src/client/index.ts
@@ -51,7 +51,7 @@ import * as ComAtprotoServerDefs from './types/com/atproto/server/defs'
 import * as ComAtprotoServerDeleteAccount from './types/com/atproto/server/deleteAccount'
 import * as ComAtprotoServerDeleteSession from './types/com/atproto/server/deleteSession'
 import * as ComAtprotoServerDescribeServer from './types/com/atproto/server/describeServer'
-import * as ComAtprotoServerGenerateSigningKey from './types/com/atproto/server/generateSigningKey'
+import * as ComAtprotoServerGetSigningKey from './types/com/atproto/server/getSigningKey'
 import * as ComAtprotoServerGetAccountInviteCodes from './types/com/atproto/server/getAccountInviteCodes'
 import * as ComAtprotoServerGetSession from './types/com/atproto/server/getSession'
 import * as ComAtprotoServerListAppPasswords from './types/com/atproto/server/listAppPasswords'
@@ -185,7 +185,7 @@ export * as ComAtprotoServerDefs from './types/com/atproto/server/defs'
 export * as ComAtprotoServerDeleteAccount from './types/com/atproto/server/deleteAccount'
 export * as ComAtprotoServerDeleteSession from './types/com/atproto/server/deleteSession'
 export * as ComAtprotoServerDescribeServer from './types/com/atproto/server/describeServer'
-export * as ComAtprotoServerGenerateSigningKey from './types/com/atproto/server/generateSigningKey'
+export * as ComAtprotoServerGetSigningKey from './types/com/atproto/server/getSigningKey'
 export * as ComAtprotoServerGetAccountInviteCodes from './types/com/atproto/server/getAccountInviteCodes'
 export * as ComAtprotoServerGetSession from './types/com/atproto/server/getSession'
 export * as ComAtprotoServerListAppPasswords from './types/com/atproto/server/listAppPasswords'
@@ -821,14 +821,14 @@ export class ServerNS {
       })
   }
 
-  generateSigningKey(
-    data?: ComAtprotoServerGenerateSigningKey.InputSchema,
-    opts?: ComAtprotoServerGenerateSigningKey.CallOptions,
-  ): Promise<ComAtprotoServerGenerateSigningKey.Response> {
+  getSigningKey(
+    data?: ComAtprotoServerGetSigningKey.InputSchema,
+    opts?: ComAtprotoServerGetSigningKey.CallOptions,
+  ): Promise<ComAtprotoServerGetSigningKey.Response> {
     return this._service.xrpc
-      .call('com.atproto.server.generateSigningKey', opts?.qp, data, opts)
+      .call('com.atproto.server.getSigningKey', opts?.qp, data, opts)
       .catch((e) => {
-        throw ComAtprotoServerGenerateSigningKey.toKnownErr(e)
+        throw ComAtprotoServerGetSigningKey.toKnownErr(e)
       })
   }
